Add RSS feed for blog
The feed was apparently already being generated, so I just made it visible (and
suppressed creating a feed for the home page because it doesn't make sense how
we're using it.)

As part of making it visible, I added social media links to the footer. We have
the RSS and YouTube links to add and for the purposes of improving spacing, I
moved GitHub from the text links to be a social icon link. The footer links are
largely duplicating the header links anyway, so it doesn't feel like a real loss
in usability.

Fixes #36

Signed-off-by: Ben Cotton <ben@kusari.dev>