Make `include_footer` default to true
We set this on each page, so it might as well be the default. If there are
specific pages we don't want the footer on, it can bet set to false on that
page's frontmatter.

I removed the explicit setting from the main content pages to make it clear
that they follow the default. I did not remove it from the blog entries because
they will not likely be edited again (although I'm open to argument that we
should remove it there for consistency and in case we decide to change the site
default in the future.)

Fixes #51

Signed-off-by: Ben Cotton <ben@kusari.dev>